Checking for valid moves
done a basic version of valid moves, where it just checks if the cell next door is the same as whe you just entered. Obviously not forrect for sudoku but it has some basic navigation of the array.

Now to work out some maths to get row, col and 3x3 based on the cell's ID.