More style, and checking for correctness
More prettiness styling, hovering over an invalid move shows that it is invalid straight away.
The hover effect has bee reworked to still show over non-editable (making them lighter grey) cells and invalid cells (making them a darker red).

Working on the logic that checks other cells to see if they are still invalid. Because there is a isValid function, I can call it on a cell where there is no longer a clash, and ask it if it is still an invalid move, although in a recent test I was able to make the code infinatly loop over the same 2 cells. To fix this I'll just have to remember which cells I'm checking (cells incase there's more then 2, sonce before I only assumed there would be 1), and skip them. for this we'll probably jsut keep using arrays since there's isn't a heaps worth of data to be checking.