Using Maps, Refactoring and Squashing Bugs.
Just a normal Friday.

So, making the code look neater, and reducing the amount of copy-pasted code. This has helped to tweak algorithims and make some of the changes to the logic that I needed to make the continual validation of cells work.

Now as the user goes through the puzzle if they decide to keep errors in hopes that they'll magically work it out later, it will search the board to see if the changes they make resolve previous issues.
Using a Map for this, since generally speaking it would be quite a sparse mapping, although in it's worst case, it could hold a whole board's worth of data.